# docs/swagger/swagger.yaml
basePath: /api/v1/
definitions:
  models.Account:
    properties:
      avatar:
        type: string
      city:
        type: string
      created_at:
        type: string
      email:
        type: string
      introduce:
        type: string
      is_root:
        type: integer
      mobile:
        type: string
      name:
        type: string
      open_id:
        type: string
      uid:
        type: integer
      union_id:
        type: string
      updated_at:
        type: string
    type: object
  utils.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server Petstore server.
  license: {}
  title: Swagger Example API
  version: 0.0.1
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: get accounts
      parameters:
      - description: name search by q
        format: email
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Account'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      security:
      - Bearer: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: add by json account
      parameters:
      - description: Add account
        in: body
        name: account
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      security:
      - Bearer: []
      summary: Add a account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete by account ID
      parameters:
      - description: Account ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.Account'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      security:
      - Bearer: []
      summary: 修改用户
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      security:
      - Bearer: []
      summary: Show a account
      tags:
      - accounts
    patch:
      consumes:
      - application/json
      description: Update by json account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update account
        in: body
        name: account
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      security:
      - Bearer: []
      summary: Update a account
      tags:
      - accounts
  /accounts/token:
    post:
      consumes:
      - application/json
      description: get token
      parameters:
      - description: Add account
        in: body
        name: account
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/string'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPError'
            type: object
      summary: 微信授权获取token接口
      tags:
      - accounts
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"